Moving Parser and target into map
